+++ src/Cdn/Cdn.Autorest/test/Get-AzFrontDoorCdnRuleSet.Tests.ps1
@@ -14,26 +14,33 @@ if(($null -eq $TestName) -or ($TestName -contains 'Get-AzFrontDoorCdnRuleSet'))
   . ($mockingPath | Select-Object -First 1).FullName
 }
 
-Describe 'Get-AzFrontDoorCdnRuleSet'  {
+Describe 'Get-AzFrontDoorCdnRuleSet' {
     BeforeAll {
-        $rulesetName = 'rsName030'
-        New-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $rulesetName
+        $script:rsName = 'rsNameGet'
+        New-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $script:rsName | Out-Null
+    }
+
+    AfterAll {
+        Remove-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $script:rsName -ErrorAction SilentlyContinue
     }
 
     It 'List' {
-        $rulesets = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName
-        $rulesets.Count | Should -BeGreaterOrEqual 1
-}
+        $rss = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName
+        $rss.Count | Should -BeGreaterOrEqual 1
+    }
 
     It 'Get' {
-        $ruleset = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $rulesetName
-        $ruleset.Name | Should -Be $rulesetName
+        $rs = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $script:rsName
+        $rs.Name | Should -Be $script:rsName
     }
 
     It 'GetViaIdentity' {
-        $rulesetObject = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $rulesetName
-        $ruleset = Get-AzFrontDoorCdnRuleSet -InputObject $rulesetObject
-        
-        $ruleset.Name | Should -Be $rulesetName
+        $rs = Get-AzFrontDoorCdnRuleSet -ProfileName $env.FrontDoorCdnProfileName -ResourceGroupName $env.ResourceGroupName -Name $script:rsName
+        $rs2 = Get-AzFrontDoorCdnRuleSet -InputObject $rs
+        $rs2.Name | Should -Be $script:rsName
+    }
+
+    It 'GetViaIdentityProfile' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
     }
 }