+++ src/Cdn/Cdn.Autorest/test/Test-AzCdnProfileMigrationCompatibility.Tests.ps1
@@ -14,11 +14,9 @@ if(($null -eq $TestName) -or ($TestName -contains 'Test-AzCdnProfileMigrationCom
   . ($mockingPath | Select-Object -First 1).FullName
 }
 
-# Playback only
 Describe 'Test-AzCdnProfileMigrationCompatibility' {
-    It 'Can' {
-        $subId = $env.SubscriptionId
-        Test-AzCdnProfileMigrationCompatibility -Subscription $subId -ProfileName cli-test-profile -ResourceGroupName cli-test-rg
+    It 'Can' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
     }
 
     It 'CanViaIdentity' -skip {